[dvsim] Share the job limit policy across launchers
The run timeout, the license wait detection and the memory ceiling all lived in
`LocalLauncher`, so a job dispatched through Slurm was subject to none of them:
it ran until the batch system's own time limit and reported no wall clock time
at all.

`Launcher` now owns that policy and the state it needs, and exposes
`_tick_runtime` and `_limit_exceeded` for a launcher's `poll` to call, with
`_kill` abstract so that a limit can stop a job and be recorded as the reason.
`_peak_mem_mb` reports nothing where a running job cannot be measured, as it
cannot be for a job on another machine, so `max_job_mem_gb` is skipped there
rather than misreported.
